cpp: throw from transaction::automatic
In C++17 it is possible to control when exceptions are thrown in the
destructor and not trigger the std::terminate call.

This enables the destructor to throw an exception if the transaction
failed without an active exception. Calling last_tx_error is now
unnecessary to check if the transaction aborted.